Move learn question to a separate component
